update: cambio en los dtos, interface y servicio de tareas. Agregado servicio para usar Token.

+++ GestionTareas_BusinessLayer/Dto/Tarea/TareaDTO.cs
@@ -2,12 +2,12 @@
 {
     public class TareaDTO
     {
-        public long Id { get; set; }        
+        public int Id { get; set; }        
         public string Descripcion { get; set; }       
         public bool IsTerminada { get; set; }
         public DateTime FechaCreacion { get; set; }
         
-        public long UsuarioId { get; set; }
+        public int UsuarioId { get; set; }
         public string Username { get; set; }
     }
 }
